pyproject.toml
[tool.poetry]
name = "playstation_store_2020_oct_scrape"
version = "0.1.0"
description = ""
authors = ["Mark Grandi <markgrandi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
arrow = "^0.17.0"
attrs = "^20.2.0"
logging_tree = "^1.8.1"
requests = "^2.24.0"
beautifulsoup4 = "^4.9.3"
warcio = "^1.7.4"
python-digitalocean = "^1.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
